+++ src/types/temperature-sensor.spec.ts
@@ -32,7 +32,7 @@ const config: PluginConfig = {
 
 const log = new Log(console, true);
 
-const hap = new Hap(socketMock, log, '031-45-154', config);
+const hap = new Hap(socketMock, log, '031-45-154', config, {});
 
 const temperatureSensor = new TemperatureSensor(hap);
 // https://developers.home.google.com/cloud-to-cloud/intents/sync